pyproject.toml
[project]
name = "agent-ia-gemini-mcp"
version = "0.1.0"
description = "Agent IA complet avec Gemini, MCP et PocketFlow"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=0.3.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    # CPU-only versions to avoid torch compatibility issues
    "faiss-cpu>=1.8.0",
    # "sentence-transformers>=3.0.0",  # Removed to avoid torch dependency
    "networkx>=3.2.0",
    "loguru>=0.7.0",
    "tenacity>=8.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
# Optional memory dependencies (install only if needed)
memory = [
    "torch>=2.0.0,<2.5.0",  # More compatible version range
    "transformers>=4.30.0",
    "sentence-transformers>=3.0.0",  # Now optional
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
